requirement.txt
discord.py
flask
pytz
pyopenssl
